+++ lib/std/jshttpcore.nim
@@ -1,0 +1,4 @@
+## Exports [jsformdata](jsformdata.html) and [jsheaders](jsheaders.html).
+
+import std/[jsformdata, jsheaders]
+export jsformdata, jsheaders